+++ ors-engine/src/main/java/org/heigit/ors/routing/graphhopper/extensions/ORSOSMReader.java
@@ -374,18 +374,8 @@ public class ORSOSMReader extends OSMReader {
         try {
             Map<Integer, Map<String, String>> tags = new HashMap<>();
             if (processNodeTags) {
-                // If we are processing the node tags then we need to obtain the tags for nodes that are on the way. We
-                // should store the internal node id though rather than the osm node as during the edge processing, we
-                // do not know the osm node id
-
-                LongArrayList osmNodeIds = way.getNodes();
-                int size = osmNodeIds.size();
-
-                for (int i = 0; i < size; i++) {
-                    // find the node
-                    long osmId = osmNodeIds.get(i);
-                    // replace the osm id with the internal id
-                    int internalId = nodeData.getId(osmId);
+                for (int nodeId : new int[]{edge.getBaseNode(), edge.getAdjNode()}) {
+                    long osmId = nodeData.getOsmId(nodeData.towerNodeToId(nodeId));
                     Map<String, String> tagsForNode = nodeTags.get(osmId);
 
                     if (countries != null && countries.containsKey(osmId)) {
@@ -395,7 +385,7 @@ public class ORSOSMReader extends OSMReader {
                     }
 
                     if (tagsForNode != null) {
-                        tags.put(internalId, tagsForNode);
+                        tags.put(nodeId, tagsForNode);
                     }
                 }
             }