+++ ors-engine/src/main/java/com/graphhopper/routing/ev/WheelchairSuitable.java
@@ -1,0 +1,17 @@
+package com.graphhopper.routing.ev;
+
+/**
+ * EncodedValue for wheelchair suitable attribute.
+ */
+public class WheelchairSuitable {
+    private WheelchairSuitable() {
+        // do not instantiate
+    }
+
+    public static final String KEY = "wheelchair_suitable";
+
+    public static SimpleBooleanEncodedValue create() {
+        return new SimpleBooleanEncodedValue(KEY);
+    }
+}
+