+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Releasing is documented in RELEASE.md
 - handle CSV parsing exceptions in CsvGraphStorageBuilder gracefully ([#2312](https://github.com/GIScience/openrouteservice/pull/2312))
 - handle the hierarchy of access restrictions properly in `AccessRestrictionsParser` ([#2311](https://github.com/GIScience/openrouteservice/pull/2311))
 - correctly resolve country borders from an PBF enriched with country tags ([#2314](https://github.com/GIScience/openrouteservice/pull/2314))
+- use direct mapping of node IDs to OSM IDs to correctly resolve node tags for edges containing cloned nodes ([#2319](https://github.com/GIScience/openrouteservice/pull/2319))
 
 ### Security
 - update postcss to 8.5.12