+++ CHANGELOG.md
@@ -58,7 +58,7 @@ Releasing is documented in RELEASE.md
 - update org.postgresql:postgresql to 42.7.11 due to [CVE-2025-49146](https://www.cve.org/CVERecord?id=CVE-2025-49146) and [CVE-2026-42198](https://www.cve.org/CVERecord?id=CVE-2026-42198)
 - update mermaid to 11.15.0
 - update uuid to 14.0.0 due to [CVE-2026-41907](https://www.cve.org/CVERecord?id=CVE-2026-41907) and [CVE-2026-41988](https://www.cve.org/CVERecord?id=CVE-2026-41988)
-- update dompurify to 3.4.11
+- update dompurify to 3.4.12
 - update spring-boot-parent to 3.5.15
 - update jackson-databind to 2.18.9 due to [CVE-2026-54515](https://www.cve.org/CVERecord?id=CVE-2026-54515)